feat : 변수 및 매핑 오류 수정

+++ src/main/java/com/bit/envdev/dto/PaymentDTO.java
@@ -16,7 +16,7 @@ public class PaymentDTO {
     private Date paymentDate;
     private String paymentUniqueNo;
     private long cartId;
-    private long contentsId;
+    private int contentsId;
     private long memberId;
 
     public Payment toEntity() {
